fixed login and sign up bug
